[Chore] 컴포넌트파일 관리를 위해 Team폴더 생성 및 경로수정 - rank폴더에서 전부 관리되던 컴포넌트파일들이 제대로 정리되지 않아   Team 폴더를 생성 후 따로 넣어 관리하게되었습니다. - 해당작업을 수행하면서 renamed처리된 파일들을 스테이징 & 커밋하였습니다.
