+++ src/components/rank/Team/SeasonTeamRankTable.tsx
@@ -3,7 +3,7 @@ import {
   useReactTable,
   ColumnDef,
 } from "@tanstack/react-table";
-import { TGameSeasonTeamRecord } from "../../types/types";
+import { TGameSeasonTeamRecord } from "../../../types/types";
 
 const SeasonTeamRankTable: React.FC<{ recordList: TGameSeasonTeamRecord }> = ({
   recordList,
@@ -82,13 +82,13 @@ const SeasonTeamRankTable: React.FC<{ recordList: TGameSeasonTeamRecord }> = ({
   });
 
   return (
-    <div className='text-xs text-center border-t-2 border-t-[#DC2626]'>
-      <table className='w-full border border-gray-200'>
+    <div className="text-xs text-center border-t-2 border-t-[#DC2626]">
+      <table className="w-full border border-gray-200">
         <thead>
           {table.getHeaderGroups().map((headerGroup) => (
-            <tr key={headerGroup.id} className='bg-gray-100'>
+            <tr key={headerGroup.id} className="bg-gray-100">
               {headerGroup.headers.map((header) => (
-                <th key={header.id} className='border py-2'>
+                <th key={header.id} className="border py-2">
                   {typeof header.column.columnDef.header === "function"
                     ? header.column.columnDef.header(header.getContext())
                     : header.column.columnDef.header}
@@ -109,7 +109,7 @@ const SeasonTeamRankTable: React.FC<{ recordList: TGameSeasonTeamRecord }> = ({
               }
             >
               {row.getVisibleCells().map((cell) => (
-                <td key={cell.id} className='border py-2'>
+                <td key={cell.id} className="border py-2">
                   {String(cell.getValue())}{" "}
                 </td>
               ))}
